+++ NEWS.md
@@ -16,6 +16,7 @@
 - add test mode to postProcess step
 - distinguish cohort table by test mode
 - add preflightChecklist to execute. Prior to a production run check all validation steps are good then execute
+- add toggle to manifests to suppress output
 
 # picard 0.0.2